test(facture): rendu PDF/HTML dans une suite dédiée, supprime test_ui
La Facture n'est pas une projection contractuelle (les documents
contractuels ont déjà test_documents_contractuels.py) : les deux tests de
rendu à valeur de test_ui.py (PDF + aperçu HTML de la facture d'énergie)
migrent vers test_facture_document.py, dédiée. Le reste de test_ui.py
(vues web génériques, rapport souscription_contrat jamais implémenté,
stub JS) n'apportait aucune preuve réelle et disparaît avec le fichier.

Corrige au passage une assertion casse-jamais-vraie héritée de test_ui
(needle capitalisée comparée à un texte .lower()).

Met à jour FEATURES.md (REQ-SOU-06, REQ-FAC-10) vers les nouvelles preuves.
